Movie, Genre 코드수정 수정 및 설명추가

+++ src/main/java/org/example/movie_api/repository/GenreRepository.java
@@ -3,5 +3,5 @@ package org.example.movie_api.repository;
 import org.example.movie_api.entity.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenreRepository extends JpaRepository<Genre, Integer> {
+public interface GenreRepository extends JpaRepository<Genre, Long> {
 }
